requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.1
google-auth==2.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3
